Login function

+++ src/main/java/com/roki/fashionfeed/web/FeedLikefullApiController.java
@@ -1,6 +1,9 @@
 package com.roki.fashionfeed.web;
 
+import com.roki.fashionfeed.config.auth.LoginUser;
+import com.roki.fashionfeed.config.auth.dto.SessionUser;
 import com.roki.fashionfeed.service.LikefullService;
+import com.roki.fashionfeed.service.UserService;
 import com.roki.fashionfeed.web.dto.LikefullSaveRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -8,31 +11,26 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
-
 @RequiredArgsConstructor
 @RestController
 public class FeedLikefullApiController {
     private final LikefullService likefullService;
+    private final UserService userService;
 
     @PutMapping("/api/{feedId}/likes")
-    public Long save(@PathVariable Long feedId, HttpServletRequest request) {
-        HttpSession session = request.getSession();
-        Object objUserId = session.getAttribute("sessionUser");
-        if (objUserId != null) {
-            Long userId = (long) (int) objUserId;
+    public Long save(@PathVariable Long feedId, @LoginUser SessionUser user) {
+        if (user != null) {
+            Long userId = userService.findUserIdByEmail(user.getEmail());
             return likefullService.save(feedId, new LikefullSaveRequestDto(userId));
         }
-        // 유저 아이디가 없는 경우 발생
         return 0L;
     }
 
-    @DeleteMapping("/api/{feedId}/likes")
-    public Long delete(@PathVariable Long feedId, HttpServletRequest request) {
-        HttpSession session = request.getSession();
-        Long userId = (long) (int) session.getAttribute("sessionUser");
-        likefullService.delete(feedId, userId);
-        return userId;
+    @DeleteMapping("/api/{feedId}/likes") //todo: feeds 가 {feedId} 앞에 들어가야하지 않을까?
+    public void delete(@PathVariable Long feedId, @LoginUser SessionUser user) {
+        if (user != null) {
+            Long userId = userService.findUserIdByEmail(user.getEmail());
+            likefullService.delete(feedId, userId);
+        }
     }
 }
